pyproject.toml
[project]
name = "langchain"
version = "0.1.0"
description = ""
authors = [
    {name = "arushii09",email = "arushiii323@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "google-genai (>=1.68.0,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
